Updated RSPS Code to be more efficient
I merged the two dictionaries used within the determine_winner function together, and updated the decision logic to reflect this change by having the win conditions within the first value, and the corresponding win messages in the second value within the key value pair.

I also updated the arguments for several functions to accept a list of inputs instead of a predetermined list of inputs making the code more scalable.

I added a print_winner function and removed the print messages from the determine_winner function to restore sole responsibility in the code.

Finally, I adopted a functional programming mindset to try and limit the number of side effects and temporary variables within the code. Best example of this is in the random_selector function.