Updated RPSLS.py
Made some further changes to the dictionary within the code. I updated the dictionary to now be a nested dictionary where each option that the user & cpu can choose has nested dictionaries within them for each win condition & win message. This makes it much easier to search the dictionaries and return the win message. On top of this, this change makes the code much easier to scale up for adding new options to the game, as well as adding more win conditions. 

I also moved the dictionary out of the determine_winner function further increasing the reusability of the code.